fixup! refactor: SupplierClient as ProofClient
